+++ lib/hexapdf/type/annotations/markup_annotation.rb
@@ -1,0 +1,70 @@
+# -*- encoding: utf-8 -*-
+#
+#--
+# This file is part of HexaPDF.
+#
+# HexaPDF - A Versatile PDF Creation and Manipulation Library For Ruby
+# Copyright (C) 2014-2017 Thomas Leitner
+#
+# HexaPDF is free software: you can redistribute it and/or modify it
+# under the terms of the GNU Affero General Public License version 3 as
+# published by the Free Software Foundation with the addition of the
+# following permission added to Section 15 as permitted in Section 7(a):
+# FOR ANY PART OF THE COVERED WORK IN WHICH THE COPYRIGHT IS OWNED BY
+# THOMAS LEITNER, THOMAS LEITNER DISCLAIMS THE WARRANTY OF NON
+# INFRINGEMENT OF THIRD PARTY RIGHTS.
+#
+# HexaPDF is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public
+# License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with HexaPDF. If not, see <http://www.gnu.org/licenses/>.
+#
+# The interactive user interfaces in modified source and object code
+# versions of HexaPDF must display Appropriate Legal Notices, as required
+# under Section 5 of the GNU Affero General Public License version 3.
+#
+# In accordance with Section 7(b) of the GNU Affero General Public
+# License, a covered work must retain the producer line in every PDF that
+# is created or manipulated using HexaPDF.
+#++
+
+require 'hexapdf/type/annotation'
+
+module HexaPDF
+  module Type
+    module Annotations
+
+      # Markup annotations are used to "mark up" a PDF document, most of the available PDF
+      # annotations are actually markup annotations.
+      #
+      # See: PDF1.7 s12.5.6.2, HexaPDF::Type::Annotation
+      class MarkupAnnotation < Annotation
+
+        define_field :T,            type: String, version: '1.1'
+        define_field :Popup,        type: :Annotation, version: '1.3'
+        define_field :CA,           type: Numeric, default: 1.0, version: '1.4'
+        define_field :RC,           type: [String, Stream], version: '1.5'
+        define_field :CreationDate, type: PDFDate, version: '1.5'
+        define_field :IRT,          type: Dictionary, version: '1.5'
+        define_field :Subj,         type: String, version: '1.5'
+        define_field :RT,           type: Symbol, default: :R, version: '1.6'
+        define_field :IT,           type: Symbol, version: '1.6'
+        define_field :ExData,       type: Dictionary, version: '1.7'
+
+        private
+
+        def perform_validation #:nodoc:
+          super
+          if key?(:RT) && !key?(:IRT)
+            yield("/IRT required if /RT field is set")
+          end
+        end
+
+      end
+
+    end
+  end
+end